Refuse a bad grammar where it is authored, not where it is run
A grammar is content, so the authoring rule holds at load as well as at
lint. `register_bundles` ran the content gate over ships, scenarios and
campaigns and did nothing to a grammar but insert it, and `content lint`
walks `assets/` offline, where a cross-mod reference cannot be decided at
all. A grammar loop beside the ship loop closes that half.

The gate itself was blind to two things. It floored each grid axis at 3
and stopped, while the generator refused three further shapes it could
not see: a stern drive that does not fit, a bow gun wider than the keel
column, and two seeded ends with no keel between them. Those are facts
about the prototypes a grammar names, and the lint had the footprints
all along. It also bounded the grid only from below, so a 2048x2048x1024
grid linted clean and multiplied to exactly 2^32 - a panic in a dev
build, a wrap to zero in a release one. `GrammarGrid::cells()` is a u64
product now, and MAX_GRAMMAR_CELLS is refused first.

92 lines of gate had no test, which is how the redundant stern arm
survived landing. Seven now, one per failure mode. `nova_wfc`'s own
refusal table asserted only `is_err()` across eight bent grammars; each
row carries the line it must be refused with, and the ninth is the
oversized grid.

Selection is NOT fixed: every caller passes `standard_hull`, so a mod's
grammar under a new id merges and is reached by nothing. Building the
picker is a feature. What changes here is the claims - the type, the
`name` field, `get_grammar`, the changelog, and the two creator pages
now say plainly that `standard_hull` is the only id anything asks for.
The grammar page's worked example declared an id nothing could reach and
seeded `railgun_section`, which no catalog holds; both are corrected.

+++ crates/nova_wfc/src/grid.rs
@@ -51,8 +51,12 @@ impl Grid {
         }
     }
 
+    /// Widened before multiplying, not after: the three sizes are authored
+    /// content, and a `u32` product of them wraps long before a `usize` one
+    /// does. `runnable` refuses a grid this large, and this is the arithmetic
+    /// that lets it be refused instead of silently becoming a grid of nothing.
     pub(crate) fn cells(&self) -> usize {
-        (self.size.x * self.size.y * self.size.z) as usize
+        self.size.x as usize * self.size.y as usize * self.size.z as usize
     }
 
     pub(crate) fn index(&self, x: usize, y: usize, z: usize) -> usize {
